Fixed leaving items behind Some code improvements
